+++ src/pools/pool.ts
@@ -111,7 +111,7 @@ export interface TasksQueueOptions {
    */
   readonly concurrency?: number
   /**
-   * Whether to enable task stealing on empty queue.
+   * Whether to enable task stealing on idle.
    *
    * @defaultValue true
    */